Move things around 11 / 12
Put ImportHook implementations in separate ImportHooks object

Also move a few utility used by it to the package object of ammonite.util, to be moved to the ammonite-api module too
