+++ amm/src/main/scala/ammonite/interp/Storage.scala
@@ -2,8 +2,7 @@ package ammonite.interp
 
 import acyclic.file
 import ammonite.ops._
-import ammonite.util.Parsers.ImportTree
-import ammonite.util.{ClassFiles, Imports, Parsers, StableRef, Timer}
+import ammonite.util.{ClassFiles, Imports, ImportTree, Parsers, StableRef, Timer}
 import ammonite.util.Codecs._
 import ammonite.util.Util.{CacheOutput, CompileCache, IvyMap}
 import org.apache.ivy.plugins.resolver.RepositoryResolver
